# main/api-docs/swagger.yaml
swagger: "2.0"
info:
  description: "Ciname API Documentation"
  version: "1.0.0"
  title: "Swagger Cinama"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/"
...
schemes:
- "http"
- "https"
paths:
...
  /pages/{id}:
    get:
      tags:
      - "pages"
      summary: "Get page by ID"
      description: ""
      operationId: "getPageById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The id that needs to be fetched. Use page1 for testing. "
        required: true
        type: "string"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Page"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Page not found"
    delete:
      tags:
      - "pages"
      summary: "Delete page by ID"
      description: ""
      operationId: "deletePage"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The ID that needs to be deleted"
        required: true
        type: "string"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "User not found"
...
  /movies/{id}:
    get:
      tags:
      - "movies"
      summary: "Get movie by ID"
      description: ""
      operationId: "getMovieById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The ID that needs to be fetched. Use movie1 for testing. "
        required: true
        type: "string"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Movie"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Movie not found"
    delete:
      tags:
      - "movies"
      summary: "Delete movie by ID"
      description: ""
      operationId: "deleteMovie"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The ID that needs to be deleted"
        required: true
        type: "string"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "Movie not found"
...
  /cinemas/{id}:
    get:
      tags:
      - "cinemas"
      summary: "Get cinema by ID"
      description: ""
      operationId: "getCinemaById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The ID that needs to be fetched. Use movie1 for testing. "
        required: true
        type: "string"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Cinema"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Cinema not found"
    delete:
      tags:
      - "cinemas"
      summary: "Delete cinema by ID"
      description: ""
      operationId: "deleteCinema"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The ID that needs to be deleted"
        required: true
        type: "string"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "Cinema not found"
...
  /cinemaBranches/{id}:
    get:
      tags:
      - "cinemaBranches"
      summary: "Get cinema by ID"
      description: ""
      operationId: "getCinemaBranchById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The ID that needs to be fetched. Use branch1 for testing. "
        required: true
        type: "string"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/CinemaBranch"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Cinema branch not found"
    delete:
      tags:
      - "cinemaBranches"
      summary: "Delete cinema branch by ID"
      description: ""
      operationId: "deleteCinemaBranch"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The ID that needs to be deleted"
        required: true
        type: "string"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "Cinema branch not found"
...
  /rooms/{id}:
    get:
      tags:
      - "rooms"
      summary: "Get room by ID"
      description: ""
      operationId: "getRoomyId"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The ID that needs to be fetched. Use room for testing. "
        required: true
        type: "string"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Room"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Room not found"
    delete:
      tags:
      - "rooms"
      summary: "Delete room by ID"
      description: ""
      operationId: "deleteRoom"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The ID that needs to be deleted"
        required: true
        type: "string"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "Room not found"
...
  /showTime/{id}:
    get:
      tags:
      - "showTime"
      summary: "Get showtime by ID"
      description: ""
      operationId: "getShowTimeById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The ID that needs to be fetched. Use room for testing. "
        required: true
        type: "string"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/ShowTime"
        400:
          description: "Invalid ID supplied"
        404:
          description: "ShowTime not found"
    delete:
      tags:
      - "showTime"
      summary: "Delete ShowTime by ID"
      description: ""
      operationId: "deleteShowTime"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The ID that needs to be deleted"
        required: true
        type: "string"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "ShowTime not found"
securityDefinitions:
  cinema_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "Modify entity in the system"
      read:pets: "View entity in the system"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  Page:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      content:
        type: "string"
      active:
        type: "boolean"
      order:
        type: "integer"
        format: "int64"
      user_id:
        type: "integer"
        format: "int64"
      parent_id:
        type: "integer"
        format: "int64"
  Movie:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      duration:
        type: "integer"
        format: "int64"
      type:
        type: "string"
      published_date:
        type: "integer"
        format: "int64"
      director:
        type: "string"
      category:
        type: "string"
      actors:
        type: "string"
      made_in:
        type: "string"
      introduction:
        type: "string"
      created_by:
        type: "integer"
        format: "int64"
  Cinema:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      icon:
        type: "string"
  CinemaBranch:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      address:
        type: "string"
      image:
        type: "string"
      cinema_id:
        type: "integer"
        format: "int64"
  Room:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      chairs:
        type: "string"
      cinema_branch_id:
        type: "integer"
        format: "int64"
      showTimes:
        type: "array"
        items:
          $ref: "#/definitions/ShowTime"
  ShowTime:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      time:
        type: "integer"
        format: "int64"
      movie_id:
        type: "integer"
        format: "int64"
...